pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pundora"
version = "1.0.0"
description = "AI-powered Dad Joke generator with text and voice modes"
authors = [{name = "Mehmet T. AKALIN", email = "mehmet@example.com"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "openai>=1.3.7",
    "elevenlabs>=0.2.26",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.2",
    "jinja2>=3.1.2",
    "aiofiles>=23.2.1",
    "qrcode[pil]==7.4.2",
    "Pillow==10.1.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.scripts]
pundora = "pundora.cli:main"
pundora-advanced = "pundora.advanced_cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/pundora"
Repository = "https://github.com/yourusername/pundora"
Issues = "https://github.com/yourusername/pundora/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["pundora*"]

[tool.setuptools.package-data]
pundora = ["templates/*.html"]